[iOS] refactor: Preview & PreviewLayer 캡쳐뷰로 옮기기
- viewDidLayoutSubviews() -> layoutSubviews()
